+++ rnalysis/data_files/r_templates/deseq2_covariate_parametric.R
@@ -1,0 +1,3 @@
+cov_res <- results(dds_res, name="$COVARIATE")
+cov_res_ordered <- cov_res[order(res$padj),]
+write.csv(as.data.frame(cov_res_ordered),file="$OUTFILE_NAME")